clean up fixed-size list array module
